Refactor isAvailable()
